+++ CHANGELOG.md
@@ -18,6 +18,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Planned
 
 ### Upcoming
+- **ADD**: Support GitHub Output Format for GitHub Actions
+- **CHANGE**: Remove `--show-all` option, use `--output-format=full` instead
 
 ### [1.7.0]
 - **CHANGE:** Bump Python minimum runtime version to 3.9.